+++ delocate/tests/test_scripts.py
@@ -402,22 +402,27 @@ def test_fix_wheel_archs(script_runner: ScriptRunner) -> None:
 def test_fuse_wheels(script_runner: ScriptRunner) -> None:
     # Some tests for wheel fusing
     with InTemporaryDirectory():
+        # Wheels need proper wheel filename for delocate-merge
+        to_wheel = "to_" + basename(PLAT_WHEEL)
+        from_wheel = "from_" + basename(PLAT_WHEEL)
         zip2dir(PLAT_WHEEL, "to_wheel")
         zip2dir(PLAT_WHEEL, "from_wheel")
-        dir2zip("to_wheel", "to_wheel.whl")
-        dir2zip("from_wheel", "from_wheel.whl")
-        script_runner.run(
-            ["delocate-fuse", "to_wheel.whl", "from_wheel.whl"], check=True
-        )
-        zip2dir("to_wheel.whl", "to_wheel_fused")
+        dir2zip("to_wheel", to_wheel)
+        dir2zip("from_wheel", from_wheel)
+        # Make sure delocate-fuse returns a non-zero exit code, it is no longer
+        # supported
+        result = script_runner.run(["delocate-fuse", to_wheel, from_wheel])
+        assert result.returncode != 0
+        script_runner.run(["delocate-merge", to_wheel, from_wheel], check=True)
+        zip2dir(to_wheel, "to_wheel_fused")
         assert_same_tree("to_wheel_fused", "from_wheel")
         # Test output argument
         os.mkdir("wheels")
         script_runner.run(
-            ["delocate-fuse", "to_wheel.whl", "from_wheel.whl", "-w", "wheels"],
+            ["delocate-merge", to_wheel, from_wheel, "-w", "wheels"],
             check=True,
         )
-        zip2dir(pjoin("wheels", "to_wheel.whl"), "to_wheel_refused")
+        zip2dir(pjoin("wheels", to_wheel), "to_wheel_refused")
         assert_same_tree("to_wheel_refused", "from_wheel")