+++ delocate/wheeltools.py
@@ -3,6 +3,8 @@
 Tools that aren't specific to delocation.
 """
 
+from __future__ import annotations
+
 import base64
 import csv
 import glob
@@ -10,6 +12,7 @@ import hashlib
 import os
 import sys
 from itertools import product
+from os import PathLike
 from os.path import abspath, basename, dirname, exists, relpath, splitext
 from os.path import join as pjoin
 from os.path import sep as psep
@@ -34,7 +37,7 @@ def _open_for_csv(name, mode):
     return open_rw(name, mode, newline="", encoding="utf-8")
 
 
-def rewrite_record(bdist_dir: str) -> None:
+def rewrite_record(bdist_dir: str | PathLike) -> None:
     """Rewrite RECORD file with hashes for all files in `wheel_sdir`.
 
     Copied from :method:`wheel.bdist_wheel.bdist_wheel.write_record`.
@@ -43,7 +46,7 @@ def rewrite_record(bdist_dir: str) -> None:
 
     Parameters
     ----------
-    bdist_dir : str
+    bdist_dir : str or Path-like
         Path of unpacked wheel file
     """
     info_dirs = glob.glob(pjoin(bdist_dir, "*.dist-info"))